requirements.txt
appdirs==1.4.0
astroid==1.4.9
autopep8==1.2.4
click==6.7
coreapi==2.1.1
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.5
django-braces==1.11.0
django-cors-middleware==1.3.1
django-oauth-toolkit==0.11.0
django-rest-framework-social-oauth2==1.0.5
django-rest-swagger==2.1.1
djangorestframework==3.5.3
dodgy==0.1.9
drfdocs==0.0.11
flake8==3.3.0
gunicorn==19.6.0
isort==4.2.5
itypes==1.1.0
lazy-object-proxy==1.2.2
mccabe==0.6.1
oauthlib==1.1.2
openapi-codec==1.2.1
packaging==16.8
pep8==1.7.0
pep8-naming==0.4.1
prospector==0.12.4
psycopg2==2.6.2
pycodestyle==2.0.0
pydocstyle==1.0.0
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.4.2
pylint==1.6.5
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.4
pyparsing==2.1.10
python-dotenv==0.6.3
python3-openid==3.0.10
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
requirements-detector==0.5.2
setoptconf==0.2.0
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
uritemplate==3.0.0
whitenoise==3.3.0
wrapt==1.10.8
yapf==0.16.0
